appendAbsence

+++ src/main/java/com/durys/jakub/leaveentitlementsservice/entilements/domain/events/LeaveEntitlementsEvent.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 
 public sealed interface LeaveEntitlementsEvent extends DomainEvent {
 
-    record AbsenceAppended(LocalDate from, LocalDate to, Integer days) implements LeaveEntitlementsEvent {}
+    record AbsenceAppended(LocalDate from, LocalDate to, Long days) implements LeaveEntitlementsEvent {}
 
     record LeaveEntitlementsGranted(LocalDate from, LocalDate to, Integer days) implements LeaveEntitlementsEvent {}
